pyproject.toml
[tool.poetry]
name = "aouth-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlalchemy = "^2.0.23"
aiomysql = "^0.2.0"
pymysql = "^1.1.0"
cryptography = "^41.0.7"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
